Reverted change to move transaction logic into analyzer: we need to begin a transaction before analysis
